functions/analyzer/requirements.txt
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
